Split _plugins.py into a directory of per-plugin modules
The monolithic _plugins.py crossed 500 lines as plugins accumulated.
Each builtin now lives in its own submodule under _plugins/, with shared
types and helpers in _core.py. The package __init__.py re-exports the
same public surface and owns the BUILTIN_PLUGINS registry, so callers
(__init__.py, _analyze.py, cli.py, tests) need no changes.

Adding the next framework preset (FastAPI, Django, Click, ...) is now a
new file plus one BUILTIN_PLUGINS entry instead of an edit deep in a
growing module.
